fix: cache MDX serialization in server utils and fix API route
Move expensive MDX serialization into getChangelogSummaries() cached
function so it runs once on cache miss instead of every request. The
page keeps connection() for CI compatibility (no DB at build time)
but now benefits from cached data AND cached MDX compilation.

Also adds connection() to the markdown API route to prevent static
rendering at build time, and removes unnecessary getCopyrightYear
cache wrapper from the layout.
