+++ src/server/utils.ts
@@ -1,6 +1,11 @@
 "use cache";
 
+import type { Element } from "hast";
 import { cacheTag } from "next/cache";
+import { serialize } from "next-mdx-remote/serialize";
+import type { Plugin } from "unified";
+import { visit } from "unist-util-visit";
+import type { ChangelogEntry } from "@/client/components/list";
 import { prismaClient } from "@/server/prisma-client";
 
 export async function getChangelogs() {
@@ -16,6 +21,42 @@ export async function getChangelogs() {
       publishedAt: "desc",
     },
   });
+}
+
+export async function getChangelogSummaries(): Promise<ChangelogEntry[]> {
+  cacheTag("changelogs");
+  const changelogs = await prismaClient.changelog.findMany({
+    include: { categories: true },
+    where: { published: true },
+    orderBy: { publishedAt: "desc" },
+  });
+
+  return Promise.all(
+    changelogs
+      .filter((changelog) => changelog.publishedAt !== null)
+      .map(async (changelog): Promise<ChangelogEntry> => {
+        const mdxSummary = await serialize(
+          changelog.summary || "",
+          {
+            mdxOptions: {
+              rehypePlugins: [
+                // @ts-expect-error
+                stripLinks,
+              ],
+            },
+          },
+          true,
+        );
+        return {
+          id: changelog.id,
+          title: changelog.title,
+          slug: changelog.slug,
+          publishedAt: new Date(changelog.publishedAt!).toUTCString(),
+          categories: changelog.categories,
+          mdxSummary,
+        };
+      }),
+  );
 }
 
 export async function getChangelog(slug: string) {
@@ -43,3 +84,17 @@ export async function getRecentChangelogs(excludeSlug: string) {
     return [];
   }
 }
+
+const stripLinks: Plugin = () => {
+  return (tree) => {
+    return visit(tree, "element", (node: Element) => {
+      if (node.tagName === "a") {
+        node.tagName = "span";
+        if (node.properties) {
+          node.properties.href = undefined;
+          node.properties.class = undefined;
+        }
+      }
+    });
+  };
+};
